feat(storybook): copy required assets to temp static directory due to 8.4 not supporting file paths
