+++ .gitignore
@@ -141,6 +141,7 @@ pkg/services/quota/quotaimpl/storage/storage.json
 
 # Ignoring frontend packages specifics
 /packages/grafana-ui/.yarn/.cache
+/packages/grafana-ui/.storybook/static
 /packages/**/dist
 /packages/**/compiled
 /packages/**/.rpt2_cache
